Refactor logging implementation
