+++ packages/logger/src/index.ts
@@ -5,7 +5,7 @@ export { Logger, LogLevel } from './logger';
 export { GlobalLogger } from './global-logger';
 export { ContextAwareLogger } from './context-aware-logger';
 export { loggerTypes } from './logger-types';
-export { registerLoggerToContainer, registerContextAwareLoggerToContainer } from './register-logger-to-container';
+export { registerLoggerToContainer } from './register-logger-to-container';
 export * from './logger-event-measurements';
 export { ConsoleLoggerClient } from './console-logger-client';
 export { LoggerEvent } from './logger-event';
