+++ packages/logger/src/register-logger-to-container.spec.ts
@@ -8,13 +8,10 @@ import * as dotenv from 'dotenv';
 import { Container } from 'inversify';
 import { AppInsightsLoggerClient } from './app-insights-logger-client';
 import { ConsoleLoggerClient } from './console-logger-client';
-import { ContextAwareAppInsightsLoggerClient } from './context-aware-app-insights-logger-client';
-import { ContextAwareConsoleLoggerClient } from './context-aware-console-logger-client';
-import { ContextAwareLogger } from './context-aware-logger';
 import { GlobalLogger } from './global-logger';
 import { LoggerClient } from './logger-client';
 import { loggerTypes } from './logger-types';
-import { registerContextAwareLoggerToContainer, registerLoggerToContainer } from './register-logger-to-container';
+import { registerLoggerToContainer } from './register-logger-to-container';
 import { OTelLoggerClient } from './otel-logger-client';
 
 /* eslint-disable @typescript-eslint/no-explicit-any */
@@ -47,41 +44,6 @@ describe('registerGlobalLoggerToContainer', () => {
         const telemetryClients = (logger as any).loggerClients as LoggerClient[];
         expect(telemetryClients.filter((c) => c instanceof AppInsightsLoggerClient)).toHaveLength(1);
         expect(telemetryClients.filter((c) => c instanceof ConsoleLoggerClient)).toHaveLength(1);
-        expect(telemetryClients.filter((c) => c instanceof OTelLoggerClient)).toHaveLength(1);
-    });
-});
-
-describe('registerContextAwareLoggerToContainer', () => {
-    beforeEach(() => {
-        container = new Container({ autoBindInjectable: true });
-    });
-
-    it('verify logger dependency resolution', () => {
-        registerLoggerToContainer(container);
-
-        verifySingletonDependencyResolution(ContextAwareAppInsightsLoggerClient);
-        verifySingletonDependencyResolution(ContextAwareConsoleLoggerClient);
-    });
-
-    it('verify context container logger resolution', () => {
-        registerLoggerToContainer(container);
-
-        const contextContainer = new Container({ autoBindInjectable: true });
-        contextContainer.parent = container;
-        registerContextAwareLoggerToContainer(contextContainer);
-
-        verifySingletonDependencyResolution(ContextAwareLogger);
-        verifySingletonDependencyResolution(ContextAwareLogger, contextContainer);
-    });
-
-    it('verify context logger resolution', () => {
-        registerLoggerToContainer(container);
-
-        const logger = container.get(ContextAwareLogger);
-
-        const telemetryClients = (logger as any).loggerClients as LoggerClient[];
-        expect(telemetryClients.filter((c) => c instanceof ContextAwareAppInsightsLoggerClient)).toHaveLength(1);
-        expect(telemetryClients.filter((c) => c instanceof ContextAwareConsoleLoggerClient)).toHaveLength(1);
         expect(telemetryClients.filter((c) => c instanceof OTelLoggerClient)).toHaveLength(1);
     });
 });
