+++ packages/web-api-send-notification-job-manager/src/send-notification-job-manager-entry-point.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 
 import { Container } from 'inversify';
-import { BaseTelemetryProperties, ContextAwareLogger } from 'logger';
+import { BaseTelemetryProperties, GlobalLogger } from 'logger';
 import { ProcessEntryPointBase } from 'service-library';
 import { SendNotificationTaskCreator } from './task/send-notification-task-creator';
 
@@ -12,7 +12,7 @@ export class SendNotificationJobManagerEntryPoint extends ProcessEntryPointBase 
     }
 
     protected async runCustomAction(container: Container): Promise<void> {
-        const logger = container.get(ContextAwareLogger);
+        const logger = container.get(GlobalLogger);
         await logger.setup();
 
         const taskCreator = container.get<SendNotificationTaskCreator>(SendNotificationTaskCreator);
