+++ packages/web-api/check-health-func/index.ts
@@ -5,7 +5,7 @@ import 'reflect-metadata';
 
 import { app, HttpRequest, HttpResponseInit, InvocationContext } from '@azure/functions';
 import { HealthCheckController } from '../src/controllers/health-check-controller';
-import { processWebRequest } from '../src/process-request';
+import { processWebRequest } from '../src/process-web-request';
 
 export async function requestHandler(request: HttpRequest, context: InvocationContext): Promise<HttpResponseInit> {
     return processWebRequest({ request, context }, HealthCheckController);
