+++ packages/web-api/src/controllers/batch-scan-result-controller.ts
@@ -4,7 +4,7 @@
 import { GuidGenerator, ServiceConfiguration } from 'common';
 import { inject, injectable } from 'inversify';
 import { isEmpty, Dictionary } from 'lodash';
-import { ContextAwareLogger } from 'logger';
+import { GlobalLogger } from 'logger';
 import {
     OnDemandPageScanRunResultProvider,
     ScanBatchRequest,
@@ -32,7 +32,7 @@ export class BatchScanResultController extends BaseScanResultController {
         @inject(ScanResponseConverter) protected readonly scanResponseConverter: ScanResponseConverter,
         @inject(GuidGenerator) protected readonly guidGenerator: GuidGenerator,
         @inject(ServiceConfiguration) protected readonly serviceConfig: ServiceConfiguration,
-        @inject(ContextAwareLogger) protected readonly logger: ContextAwareLogger,
+        @inject(GlobalLogger) protected readonly logger: GlobalLogger,
     ) {
         super(logger);
     }
