+++ packages/web-api/src/get-process-life-cycle-container.ts
@@ -4,14 +4,14 @@
 import { ApplicationInsightsClient, registerAzureServicesToContainer, secretNames, SecretProvider } from 'azure-services';
 import { IoC, setupRuntimeConfigContainer } from 'common';
 import * as inversify from 'inversify';
-import { isNil } from 'lodash';
+import { isEmpty } from 'lodash';
 import { registerLoggerToContainer } from 'logger';
 import { webApiTypeNames } from './web-api-types';
 
 let processLifeCycleContainer: inversify.Container;
 
 export function getProcessLifeCycleContainer(): inversify.Container {
-    if (isNil(processLifeCycleContainer)) {
+    if (isEmpty(processLifeCycleContainer)) {
         processLifeCycleContainer = new inversify.Container({ autoBindInjectable: true });
         setupRuntimeConfigContainer(processLifeCycleContainer);
         registerLoggerToContainer(processLifeCycleContainer);
