+++ packages/web-api/src/process-web-request.ts
@@ -3,8 +3,8 @@
 
 import { HttpResponseInit } from '@azure/functions';
 import { AppContext, getGlobalWebControllerDispatcher, Newable, WebController } from 'service-library';
+import { Container } from 'inversify';
 import { getProcessLifeCycleContainer } from './get-process-life-cycle-container';
-import { setupRequestContextIocContainer } from './setup-request-context-ioc-container';
 
 /* eslint-disable @typescript-eslint/no-explicit-any */
 
@@ -15,7 +15,9 @@ export async function processWebRequest(
 ): Promise<HttpResponseInit> {
     const processLifeCycleContainer = getProcessLifeCycleContainer();
     const dispatcher = await getGlobalWebControllerDispatcher(processLifeCycleContainer);
-    const requestContainer = setupRequestContextIocContainer(processLifeCycleContainer);
 
-    return dispatcher.processRequest(requestContainer, controllerType, appContext, ...args);
+    const container = new Container({ autoBindInjectable: true });
+    container.parent = processLifeCycleContainer;
+
+    return dispatcher.processRequest(container, controllerType, appContext, ...args);
 }
