+++ packages/web-workers/src/process-web-request.ts
@@ -4,20 +4,15 @@
 import { AppContext, getGlobalWebControllerDispatcher, Newable, WebController } from 'service-library';
 import { Container } from 'inversify';
 import { getProcessLifeCycleContainer } from './get-process-life-cycle-container';
-import { setupRequestContextIocContainer } from './setup-request-context-ioc-container';
 
 /* eslint-disable @typescript-eslint/no-explicit-any */
 
 export async function processWebRequest(appContext: AppContext, controllerType: Newable<WebController>, ...args: any[]): Promise<any> {
     const processLifeCycleContainer = getProcessLifeCycleContainer();
-    const requestContainer = getRequestContainer(processLifeCycleContainer);
     const dispatcher = await getGlobalWebControllerDispatcher(processLifeCycleContainer);
 
-    return dispatcher.processRequest(requestContainer, controllerType, appContext, ...args);
-}
+    const container = new Container({ autoBindInjectable: true });
+    container.parent = processLifeCycleContainer;
 
-export function getRequestContainer(container?: Container): Container {
-    const processLifeCycleContainer = container ?? getProcessLifeCycleContainer();
-
-    return setupRequestContextIocContainer(processLifeCycleContainer);
+    return dispatcher.processRequest(container, controllerType, appContext, ...args);
 }
